refactor(package-manager): move install phases into modules
