+++ pnpm/crates/package-manager/src/install/prepare_modules_state.rs
@@ -1,0 +1,398 @@
+use super::{
+    Arc, Catalogs, Config, HashSet, Host, IncludedDependencies, InstallError, Lockfile, LogEvent,
+    LogLevel, Modules, NodeLinker, PackageManifest, Path, PathBuf, PnpmLog, RebuildOptions,
+    Reporter, ResolutionVerifier, Stage, StageLog, SummaryLog, SystemTime, build_workspace_state,
+    check_modules_settings_diff, frozen_tree_intact, gvs_build_marker_present,
+    has_newly_allowed_ignored_builds, has_revoked_allowed_builds, map_frozen_lockfile_error,
+    modules_consistent_with, modules_layout_consistent_with, unapproved_recorded_ignored_builds,
+    update_workspace_state, verify_lockfile_eagerly,
+};
+
+pub(super) struct PrepareModulesStateInputs<'a, 'install> {
+    pub(super) resolve_only: bool,
+    pub(super) take_frozen_path: bool,
+    pub(super) config: &'static Config,
+    pub(super) filtered_install: bool,
+    pub(super) installs_only: bool,
+    pub(super) workspace_root: &'a Path,
+    pub(super) included: IncludedDependencies,
+    pub(super) current_lockfile: Option<&'a Lockfile>,
+    pub(super) requested_importer_ids: Option<&'a HashSet<String>>,
+    pub(super) node_linker: NodeLinker,
+    pub(super) disable_optimistic_repeat_install: bool,
+    pub(super) lockfile: Option<&'a Lockfile>,
+    pub(super) supported_architectures:
+        Option<&'a pacquet_package_is_installable::SupportedArchitectures>,
+    pub(super) rebuild: Option<&'a RebuildOptions>,
+    pub(super) resolution_verifiers: &'a [Arc<dyn ResolutionVerifier>],
+    pub(super) derived_lockfile_path: Option<&'a Path>,
+    pub(super) lockfile_verification_override:
+        Option<super::LockfileVerificationOverride<'install>>,
+    pub(super) lockfile_synthesized_from_current: bool,
+    pub(super) lockfile_was_fast_updated: bool,
+    pub(super) catalogs: &'a Catalogs,
+    pub(super) project_manifests: &'a [(PathBuf, &'a PackageManifest)],
+    pub(super) prefix: &'a String,
+}
+
+pub(super) struct PreparedModulesState<'install> {
+    pub(super) old_modules: Option<pacquet_modules_yaml::ModulesLayout>,
+    pub(super) previous_modules_metadata: Option<Modules>,
+    pub(super) is_inconsistent: bool,
+    pub(super) lockfile_verification_override:
+        Option<super::LockfileVerificationOverride<'install>>,
+}
+
+pub(super) async fn prepare_modules_state<'install, Reporter: self::Reporter + 'static>(
+    inputs: PrepareModulesStateInputs<'_, 'install>,
+) -> Result<Option<PreparedModulesState<'install>>, InstallError> {
+    let PrepareModulesStateInputs {
+        resolve_only,
+        take_frozen_path,
+        config,
+        filtered_install,
+        installs_only,
+        workspace_root,
+        included,
+        current_lockfile,
+        requested_importer_ids,
+        node_linker,
+        disable_optimistic_repeat_install,
+        lockfile,
+        supported_architectures,
+        rebuild,
+        resolution_verifiers,
+        derived_lockfile_path,
+        lockfile_verification_override,
+        lockfile_synthesized_from_current,
+        lockfile_was_fast_updated,
+        catalogs,
+        project_manifests,
+        prefix,
+    } = inputs;
+    // A no-op still refreshes workspace state so `verifyDepsBeforeRun`
+    // does not treat the materialized tree as stale.
+    let modules_manifest_res = if !resolve_only || take_frozen_path {
+        pacquet_modules_yaml::read_modules_layout::<Host>(&config.modules_dir)
+    } else {
+        Ok(None)
+    };
+    let read_failed = modules_manifest_res.is_err();
+    if let Err(err) = &modules_manifest_res {
+        tracing::warn!(
+            target: "pacquet::install",
+            ?err,
+            "failed to read .modules.yaml; treating as an inconsistent node_modules directory",
+        );
+    }
+    let old_modules = modules_manifest_res.ok().flatten();
+    let modules_manifest = old_modules.as_ref();
+    // A filtered install rewrites `.modules.yaml` from the selected
+    // projects' state merged over the previous file's, so losing the
+    // previous contents would drop every unselected project's entries.
+    // An unreadable *layout* is already handled as an inconsistent
+    // `node_modules` — the purge rebuilds everything and the merge is
+    // skipped — but a file whose layout parses while some later field
+    // does not would otherwise merge against `None` and silently prune
+    // those entries, so that case fails instead.
+    let previous_modules_metadata = if !resolve_only && !read_failed {
+        match pacquet_modules_yaml::read_modules_manifest::<Host>(&config.modules_dir) {
+            Ok(modules) => modules,
+            // A filtered install merges the unselected importers'
+            // entries out of this file, so it cannot proceed
+            // without it; an unfiltered install only loses the
+            // orphan hoist-link cleanup.
+            Err(error) if filtered_install => return Err(InstallError::ReadModules(error)),
+            Err(error) => {
+                tracing::warn!(
+                    target: "pacquet::install",
+                    ?error,
+                    "failed to fully parse .modules.yaml; skipping orphan hoist-link cleanup",
+                );
+                None
+            }
+        }
+    } else {
+        None
+    };
+    // The purge keys off *layout* drift only, not `included`: an
+    // included (`--prod`<->full) change is handled by relinking, so it
+    // must not wipe the user's `node_modules` contents. See
+    // [`modules_layout_consistent_with`].
+    let is_inconsistent = read_failed
+        || match &modules_manifest {
+            Some(modules) => !modules_layout_consistent_with(modules, config, node_linker),
+            // Treat existence-check errors conservatively as inconsistent.
+            None => config
+                .modules_dir
+                .join(pacquet_modules_yaml::MODULES_FILENAME)
+                .try_exists()
+                .unwrap_or(true),
+        };
+
+    if !resolve_only && is_inconsistent {
+        // A plain install may recreate the drifted modules dir;
+        // `add` / `remove` must surface the drift instead
+        // (upstream `validateModules` with `forceNewModules =
+        // installsOnly`).
+        if !installs_only && let Some(modules) = modules_manifest {
+            check_modules_settings_diff(modules, config)?;
+        }
+        let (is_safe, target_dir) = if config.modules_dir.exists() {
+            match (
+                std::fs::canonicalize(&config.modules_dir),
+                std::fs::canonicalize(workspace_root),
+            ) {
+                (Ok(modules_canon), Ok(root_canon)) => {
+                    (modules_canon.starts_with(&root_canon), Some(modules_canon))
+                }
+                _ => (false, None),
+            }
+        } else {
+            (true, None)
+        };
+        if is_safe {
+            if let Some(target) = target_dir {
+                match std::fs::read_dir(&target) {
+                    Ok(entries) => {
+                        for entry_res in entries {
+                            let entry = match entry_res {
+                                Ok(e) => e,
+                                Err(err) if err.kind() == std::io::ErrorKind::NotFound => {
+                                    continue;
+                                }
+                                Err(err) => return Err(InstallError::RemoveModulesDir(err)),
+                            };
+                            let file_name = entry.file_name();
+                            let file_name_str = file_name.to_string_lossy();
+
+                            let is_hidden = file_name_str.starts_with('.');
+                            let is_pnpm_hidden = file_name_str == ".bin"
+                                || file_name_str == ".modules.yaml"
+                                || config
+                                    .virtual_store_dir
+                                    .file_name()
+                                    .is_some_and(|n| n == file_name_str.as_ref())
+                                || modules_manifest.as_ref().is_some_and(|manifest| {
+                                    let mut old_vs =
+                                        std::path::PathBuf::from(&manifest.virtual_store_dir);
+                                    if old_vs.is_relative() {
+                                        old_vs = config.modules_dir.join(old_vs);
+                                    }
+                                    old_vs.starts_with(&config.modules_dir)
+                                        && old_vs
+                                            .file_name()
+                                            .is_some_and(|n| n == file_name_str.as_ref())
+                                });
+
+                            if is_hidden && !is_pnpm_hidden {
+                                continue;
+                            }
+
+                            if entry.file_type().is_ok_and(|t| t.is_dir()) {
+                                #[cfg(windows)]
+                                let is_removed =
+                                    pacquet_fs::remove_symlink_dir(&entry.path()).is_ok();
+                                #[cfg(not(windows))]
+                                let is_removed = false;
+
+                                if !is_removed
+                                    && let Err(err) = std::fs::remove_dir_all(entry.path())
+                                    && err.kind() != std::io::ErrorKind::NotFound
+                                {
+                                    return Err(InstallError::RemoveModulesDir(err));
+                                }
+                            } else if let Err(err) = std::fs::remove_file(entry.path())
+                                && err.kind() != std::io::ErrorKind::NotFound
+                            {
+                                return Err(InstallError::RemoveModulesDir(err));
+                            }
+                        }
+                    }
+                    Err(err) if err.kind() == std::io::ErrorKind::NotFound => {}
+                    Err(err) => return Err(InstallError::RemoveModulesDir(err)),
+                }
+            }
+        } else {
+            if filtered_install {
+                return Err(InstallError::UnsafeFilteredModulesDir {
+                    modules_dir: config.modules_dir.clone(),
+                    workspace_root: workspace_root.to_path_buf(),
+                });
+            }
+            tracing::warn!(
+                ?config.modules_dir,
+                "refusing to remove inconsistent modules directory outside the project root",
+            );
+        }
+    }
+
+    // Remove direct links from dependency groups excluded by this
+    // run. Unfiltered installs can use the global `included` value
+    // recorded in `.modules.yaml`; filtered installs may retain
+    // importers materialized with different group sets, so they
+    // conservatively prune every excluded group from only the
+    // selected workspace-link closure.
+    if !resolve_only
+        && !is_inconsistent
+        && let Some(modules) = modules_manifest
+        && let Some(current) = current_lockfile.as_ref()
+        && (filtered_install || modules.included != included)
+    {
+        let selected_prune_importer_ids = requested_importer_ids.as_ref().map(|requested| {
+            crate::materialization_closure(
+                current,
+                workspace_root,
+                requested,
+                included,
+                &crate::SkippedSnapshots::new(),
+            )
+            .importer_ids
+        });
+        let previously_included = if filtered_install {
+            IncludedDependencies {
+                dependencies: true,
+                dev_dependencies: true,
+                optional_dependencies: true,
+            }
+        } else {
+            modules.included
+        };
+        crate::prune_direct_deps_excluded_by_groups(
+            current,
+            previously_included,
+            included,
+            workspace_root,
+            config,
+            selected_prune_importer_ids.as_ref(),
+        )
+        .map_err(InstallError::PruneDirectDeps)?;
+    }
+
+    let modules_cache_prune_due = modules_manifest.as_ref().is_some_and(|modules| {
+        crate::prune_virtual_store::should_prune_virtual_store(
+            crate::prune_virtual_store::same_dir(
+                config.effective_virtual_store_dir(),
+                &config.global_virtual_store_dir,
+            ),
+            Some(modules.pruned_at.as_str()),
+            config.modules_cache_max_age,
+            SystemTime::now(),
+        )
+    });
+
+    if take_frozen_path
+            && !filtered_install
+            && !disable_optimistic_repeat_install
+            // `--force` reinstalls everything, so an up-to-date tree
+            // must not short-circuit the materialization.
+            && !config.force
+            && let Some(wanted_lockfile) = lockfile
+            && let Some(current) = current_lockfile
+            && wanted_lockfile == current
+            && let Some(modules) = modules_manifest.as_ref()
+            && modules_consistent_with(modules, config, node_linker, included)
+            // A `supportedArchitectures` change alters the skip set
+            // without touching the lockfile or `.modules.yaml`, so the
+            // unchanged-layout premise doesn't hold and the platform
+            // packages must be re-evaluated.
+            && crate::optimistic_repeat_install::recorded_supported_architectures_match(
+                workspace_root,
+                supported_architectures,
+            )
+            // An `allowBuilds` change that now permits a previously-ignored
+            // build must rebuild it, even though the lockfile and layout are
+            // unchanged.
+            && !has_newly_allowed_ignored_builds(modules, config)
+            // The mirror image: an approval the user has since withdrawn
+            // must be re-evaluated, or a strict install would exit 0 on a
+            // package it is no longer allowed to build.
+            && !has_revoked_allowed_builds(modules, config)
+            // A build marker lives in the shared slot, outside every
+            // project-state input checked above. Let materialization inspect
+            // buildable and patched GVS slots instead of declaring the local
+            // tree complete from importer links alone.
+            && !gvs_build_marker_present(wanted_lockfile, config, workspace_root)
+            // An explicit `pacquet rebuild` always re-runs the build phase,
+            // so it never short-circuits here.
+            && rebuild.is_none()
+            && !modules_cache_prune_due
+            && frozen_tree_intact(wanted_lockfile, modules, config, workspace_root, node_linker)
+    {
+        // The full frozen path runs the offline structural
+        // name gate before any materialization; the up-to-date
+        // early return must not skip it (the resolution-verifier
+        // fan-out below is policy-gated and can be empty).
+        pacquet_lockfile_verification::verify_lockfile_dependency_names(wanted_lockfile)
+            .map_err(InstallError::LockfileVerification)?;
+        // Nothing to materialize means no fetch to overlap; verify
+        // eagerly before the up-to-date early return.
+        if let Some(lockfile_verification_override) = lockfile_verification_override {
+            lockfile_verification_override.await.map_err(map_frozen_lockfile_error)?;
+        } else {
+            verify_lockfile_eagerly::<Reporter>(
+                wanted_lockfile,
+                resolution_verifiers,
+                derived_lockfile_path,
+                &config.cache_dir,
+            )
+            .await?;
+        }
+        // Keep `strictDepBuilds` enforced on the up-to-date path: a
+        // rerun after an `ERR_PNPM_IGNORED_BUILDS` failure must not
+        // exit 0 just because the lockfile and layout are unchanged.
+        // Checked after verification (a tampered lockfile fails first)
+        // and before the "up to date" log so the command doesn't
+        // claim success.
+        // `Err` (malformed `allowBuilds`) is unreachable here — the
+        // `has_newly_allowed_ignored_builds` guard above returns `true`
+        // on the same `from_config` error and skips this block — so a
+        // bad policy is surfaced by the full install instead.
+        if config.strict_dep_builds
+            && let Ok(Some(package_names)) = unapproved_recorded_ignored_builds(modules, config)
+        {
+            return Err(InstallError::IgnoredBuilds { package_names });
+        }
+        Reporter::emit(&LogEvent::Pnpm(PnpmLog {
+            level: LogLevel::Info,
+            message: "Lockfile is up to date, resolution step is skipped".to_string(),
+            prefix: prefix.clone(),
+        }));
+        Reporter::emit(&LogEvent::Stage(StageLog {
+            level: LogLevel::Debug,
+            prefix: prefix.clone(),
+            stage: Stage::ImportingDone,
+        }));
+        if (lockfile_synthesized_from_current || lockfile_was_fast_updated) && config.lockfile {
+            wanted_lockfile
+                .save_to_path(&workspace_root.join(Lockfile::FILE_NAME))
+                .map_err(InstallError::SaveWantedLockfile)?;
+        }
+        update_workspace_state(
+            workspace_root,
+            &build_workspace_state(
+                workspace_root,
+                config,
+                node_linker,
+                included,
+                supported_architectures,
+                catalogs,
+                project_manifests,
+                filtered_install,
+            ),
+        )
+        .map_err(InstallError::WriteWorkspaceState)?;
+        Reporter::emit(&LogEvent::Summary(SummaryLog {
+            level: LogLevel::Debug,
+            prefix: prefix.to_owned(),
+        }));
+        return Ok(None);
+    }
+
+    Ok(Some(PreparedModulesState {
+        old_modules,
+        previous_modules_metadata,
+        is_inconsistent,
+        lockfile_verification_override,
+    }))
+}
