+++ pnpm/crates/package-manager/src/install/run.rs
@@ -1,28 +1,19 @@
 use super::{
-    Arc, AtomicU8, AuthHeaders, BTreeMap, BTreeSet, Catalogs, Config, ContextLog, DependencyGroup,
-    DepsRequiringBuildSink, FastUpdateImporterLockfileOptions, FreshnessCheckError, HashSet,
-    HoistedDependencies, Host, InMemoryPackageMetaCache, IncludedDependencies, Install,
-    InstallError, InstallFrozenLockfile, InstallRunOptions, InstallWithFreshLockfile, IsTerminal,
-    Lockfile, LogEvent, LogLevel, MemCache, Modules, NodeLinker, OptimisticRepeatInstallCheck,
-    OptimisticRepeatInstallDecision, PackageManifest, Path, PathBuf, PeerIssuesSink, PnpmLog,
-    ProjectMutation, ProjectScriptsInputs, RebuildOptions, Reporter, ResolutionVerifier,
-    ResolvedPackages, ScopeLog, Stage, StageLog, SummaryLog, SystemTime, ThrottledClient,
-    UpdateSeedPolicy, WorkspaceInstallSelection, build_modules_manifest,
-    build_project_manifests_list, build_resolution_verifiers,
-    build_root_importer_project_manifests_list, build_selected_project_manifests_list,
-    build_workspace_packages_map, build_workspace_state, check_lockfile_freshness,
-    check_modules_settings_diff, check_optimistic_repeat_install,
-    configured_or_discovered_workspace_dir, dev_preinstall_already_ran, drain_settled_projects,
-    emit_initial_package_manifest, frozen_tree_intact, get_catalogs_from_workspace_manifest,
-    gvs_build_marker_present, gvs_build_markers_may_require_recovery,
-    has_newly_allowed_ignored_builds, has_revoked_allowed_builds, load_workspace_projects,
-    map_frozen_lockfile_error, merge_filtered_modules_metadata, merge_pending_builds,
-    modules_consistent_with, modules_layout_consistent_with, node_version_from_engines_runtime,
-    order_project_lifecycle_groups, project_requires_lifecycle_scripts,
-    projects_running_own_scripts, record_lockfile_verified, run_dev_preinstall,
-    run_projects_lifecycle_scripts, selected_manifest_freshness_inputs,
-    try_fast_update_importer_lockfile, unapproved_recorded_ignored_builds, update_workspace_state,
-    verify_lockfile_eagerly, write_modules_manifest,
+    ApplyMaterializationInputs, Arc, AtomicU8, ContextLog, DependencyGroup,
+    FastUpdateImporterLockfileOptions, FreshnessCheckError, HashSet, Host,
+    InMemoryPackageMetaCache, IncludedDependencies, Install, InstallError, InstallRunOptions,
+    IsTerminal, Lockfile, LogEvent, LogLevel, MaterializationInputs, MaterializationOutput,
+    OptimisticRepeatInstallCheck, OptimisticRepeatInstallDecision, Path, PnpmLog,
+    PrepareModulesStateInputs, PreparedModulesState, Reporter, ScopeLog, Stage, StageLog,
+    SummaryLog, UpdateSeedPolicy, apply_materialization_result, build_project_manifests_list,
+    build_resolution_verifiers, build_root_importer_project_manifests_list,
+    build_selected_project_manifests_list, check_lockfile_freshness,
+    check_optimistic_repeat_install, configured_or_discovered_workspace_dir,
+    dev_preinstall_already_ran, emit_initial_package_manifest,
+    get_catalogs_from_workspace_manifest, gvs_build_marker_present,
+    gvs_build_markers_may_require_recovery, load_workspace_projects, map_frozen_lockfile_error,
+    materialize, prepare_modules_state, run_dev_preinstall, selected_manifest_freshness_inputs,
+    try_fast_update_importer_lockfile, unapproved_recorded_ignored_builds, verify_lockfile_eagerly,
 };
 use pacquet_executor::DEV_PREINSTALL_STAGE;
 
@@ -907,1538 +898,4 @@ where
         })
         .await
     }
-}
-
-struct PrepareModulesStateInputs<'a, 'install> {
-    resolve_only: bool,
-    take_frozen_path: bool,
-    config: &'static Config,
-    filtered_install: bool,
-    installs_only: bool,
-    workspace_root: &'a Path,
-    included: IncludedDependencies,
-    current_lockfile: Option<&'a Lockfile>,
-    requested_importer_ids: Option<&'a HashSet<String>>,
-    node_linker: NodeLinker,
-    disable_optimistic_repeat_install: bool,
-    lockfile: Option<&'a Lockfile>,
-    supported_architectures: Option<&'a pacquet_package_is_installable::SupportedArchitectures>,
-    rebuild: Option<&'a RebuildOptions>,
-    resolution_verifiers: &'a [Arc<dyn ResolutionVerifier>],
-    derived_lockfile_path: Option<&'a Path>,
-    lockfile_verification_override: Option<super::LockfileVerificationOverride<'install>>,
-    lockfile_synthesized_from_current: bool,
-    lockfile_was_fast_updated: bool,
-    catalogs: &'a Catalogs,
-    project_manifests: &'a [(PathBuf, &'a PackageManifest)],
-    prefix: &'a String,
-}
-
-struct PreparedModulesState<'install> {
-    old_modules: Option<pacquet_modules_yaml::ModulesLayout>,
-    previous_modules_metadata: Option<Modules>,
-    is_inconsistent: bool,
-    lockfile_verification_override: Option<super::LockfileVerificationOverride<'install>>,
-}
-
-async fn prepare_modules_state<'install, Reporter: self::Reporter + 'static>(
-    inputs: PrepareModulesStateInputs<'_, 'install>,
-) -> Result<Option<PreparedModulesState<'install>>, InstallError> {
-    let PrepareModulesStateInputs {
-        resolve_only,
-        take_frozen_path,
-        config,
-        filtered_install,
-        installs_only,
-        workspace_root,
-        included,
-        current_lockfile,
-        requested_importer_ids,
-        node_linker,
-        disable_optimistic_repeat_install,
-        lockfile,
-        supported_architectures,
-        rebuild,
-        resolution_verifiers,
-        derived_lockfile_path,
-        lockfile_verification_override,
-        lockfile_synthesized_from_current,
-        lockfile_was_fast_updated,
-        catalogs,
-        project_manifests,
-        prefix,
-    } = inputs;
-    // A no-op still refreshes workspace state so `verifyDepsBeforeRun`
-    // does not treat the materialized tree as stale.
-    let modules_manifest_res = if !resolve_only || take_frozen_path {
-        pacquet_modules_yaml::read_modules_layout::<Host>(&config.modules_dir)
-    } else {
-        Ok(None)
-    };
-    let read_failed = modules_manifest_res.is_err();
-    if let Err(err) = &modules_manifest_res {
-        tracing::warn!(
-            target: "pacquet::install",
-            ?err,
-            "failed to read .modules.yaml; treating as an inconsistent node_modules directory",
-        );
-    }
-    let old_modules = modules_manifest_res.ok().flatten();
-    let modules_manifest = old_modules.as_ref();
-    // A filtered install rewrites `.modules.yaml` from the selected
-    // projects' state merged over the previous file's, so losing the
-    // previous contents would drop every unselected project's entries.
-    // An unreadable *layout* is already handled as an inconsistent
-    // `node_modules` — the purge rebuilds everything and the merge is
-    // skipped — but a file whose layout parses while some later field
-    // does not would otherwise merge against `None` and silently prune
-    // those entries, so that case fails instead.
-    let previous_modules_metadata = if !resolve_only && !read_failed {
-        match pacquet_modules_yaml::read_modules_manifest::<Host>(&config.modules_dir) {
-            Ok(modules) => modules,
-            // A filtered install merges the unselected importers'
-            // entries out of this file, so it cannot proceed
-            // without it; an unfiltered install only loses the
-            // orphan hoist-link cleanup.
-            Err(error) if filtered_install => return Err(InstallError::ReadModules(error)),
-            Err(error) => {
-                tracing::warn!(
-                    target: "pacquet::install",
-                    ?error,
-                    "failed to fully parse .modules.yaml; skipping orphan hoist-link cleanup",
-                );
-                None
-            }
-        }
-    } else {
-        None
-    };
-    // The purge keys off *layout* drift only, not `included`: an
-    // included (`--prod`<->full) change is handled by relinking, so it
-    // must not wipe the user's `node_modules` contents. See
-    // [`modules_layout_consistent_with`].
-    let is_inconsistent = read_failed
-        || match &modules_manifest {
-            Some(modules) => !modules_layout_consistent_with(modules, config, node_linker),
-            // Treat existence-check errors conservatively as inconsistent.
-            None => config
-                .modules_dir
-                .join(pacquet_modules_yaml::MODULES_FILENAME)
-                .try_exists()
-                .unwrap_or(true),
-        };
-
-    if !resolve_only && is_inconsistent {
-        // A plain install may recreate the drifted modules dir;
-        // `add` / `remove` must surface the drift instead
-        // (upstream `validateModules` with `forceNewModules =
-        // installsOnly`).
-        if !installs_only && let Some(modules) = modules_manifest {
-            check_modules_settings_diff(modules, config)?;
-        }
-        let (is_safe, target_dir) = if config.modules_dir.exists() {
-            match (
-                std::fs::canonicalize(&config.modules_dir),
-                std::fs::canonicalize(workspace_root),
-            ) {
-                (Ok(modules_canon), Ok(root_canon)) => {
-                    (modules_canon.starts_with(&root_canon), Some(modules_canon))
-                }
-                _ => (false, None),
-            }
-        } else {
-            (true, None)
-        };
-        if is_safe {
-            if let Some(target) = target_dir {
-                match std::fs::read_dir(&target) {
-                    Ok(entries) => {
-                        for entry_res in entries {
-                            let entry = match entry_res {
-                                Ok(e) => e,
-                                Err(err) if err.kind() == std::io::ErrorKind::NotFound => {
-                                    continue;
-                                }
-                                Err(err) => return Err(InstallError::RemoveModulesDir(err)),
-                            };
-                            let file_name = entry.file_name();
-                            let file_name_str = file_name.to_string_lossy();
-
-                            let is_hidden = file_name_str.starts_with('.');
-                            let is_pnpm_hidden = file_name_str == ".bin"
-                                || file_name_str == ".modules.yaml"
-                                || config
-                                    .virtual_store_dir
-                                    .file_name()
-                                    .is_some_and(|n| n == file_name_str.as_ref())
-                                || modules_manifest.as_ref().is_some_and(|manifest| {
-                                    let mut old_vs =
-                                        std::path::PathBuf::from(&manifest.virtual_store_dir);
-                                    if old_vs.is_relative() {
-                                        old_vs = config.modules_dir.join(old_vs);
-                                    }
-                                    old_vs.starts_with(&config.modules_dir)
-                                        && old_vs
-                                            .file_name()
-                                            .is_some_and(|n| n == file_name_str.as_ref())
-                                });
-
-                            if is_hidden && !is_pnpm_hidden {
-                                continue;
-                            }
-
-                            if entry.file_type().is_ok_and(|t| t.is_dir()) {
-                                #[cfg(windows)]
-                                let is_removed =
-                                    pacquet_fs::remove_symlink_dir(&entry.path()).is_ok();
-                                #[cfg(not(windows))]
-                                let is_removed = false;
-
-                                if !is_removed
-                                    && let Err(err) = std::fs::remove_dir_all(entry.path())
-                                    && err.kind() != std::io::ErrorKind::NotFound
-                                {
-                                    return Err(InstallError::RemoveModulesDir(err));
-                                }
-                            } else if let Err(err) = std::fs::remove_file(entry.path())
-                                && err.kind() != std::io::ErrorKind::NotFound
-                            {
-                                return Err(InstallError::RemoveModulesDir(err));
-                            }
-                        }
-                    }
-                    Err(err) if err.kind() == std::io::ErrorKind::NotFound => {}
-                    Err(err) => return Err(InstallError::RemoveModulesDir(err)),
-                }
-            }
-        } else {
-            if filtered_install {
-                return Err(InstallError::UnsafeFilteredModulesDir {
-                    modules_dir: config.modules_dir.clone(),
-                    workspace_root: workspace_root.to_path_buf(),
-                });
-            }
-            tracing::warn!(
-                ?config.modules_dir,
-                "refusing to remove inconsistent modules directory outside the project root",
-            );
-        }
-    }
-
-    // Remove direct links from dependency groups excluded by this
-    // run. Unfiltered installs can use the global `included` value
-    // recorded in `.modules.yaml`; filtered installs may retain
-    // importers materialized with different group sets, so they
-    // conservatively prune every excluded group from only the
-    // selected workspace-link closure.
-    if !resolve_only
-        && !is_inconsistent
-        && let Some(modules) = modules_manifest
-        && let Some(current) = current_lockfile.as_ref()
-        && (filtered_install || modules.included != included)
-    {
-        let selected_prune_importer_ids = requested_importer_ids.as_ref().map(|requested| {
-            crate::materialization_closure(
-                current,
-                workspace_root,
-                requested,
-                included,
-                &crate::SkippedSnapshots::new(),
-            )
-            .importer_ids
-        });
-        let previously_included = if filtered_install {
-            IncludedDependencies {
-                dependencies: true,
-                dev_dependencies: true,
-                optional_dependencies: true,
-            }
-        } else {
-            modules.included
-        };
-        crate::prune_direct_deps_excluded_by_groups(
-            current,
-            previously_included,
-            included,
-            workspace_root,
-            config,
-            selected_prune_importer_ids.as_ref(),
-        )
-        .map_err(InstallError::PruneDirectDeps)?;
-    }
-
-    let modules_cache_prune_due = modules_manifest.as_ref().is_some_and(|modules| {
-        crate::prune_virtual_store::should_prune_virtual_store(
-            crate::prune_virtual_store::same_dir(
-                config.effective_virtual_store_dir(),
-                &config.global_virtual_store_dir,
-            ),
-            Some(modules.pruned_at.as_str()),
-            config.modules_cache_max_age,
-            SystemTime::now(),
-        )
-    });
-
-    if take_frozen_path
-            && !filtered_install
-            && !disable_optimistic_repeat_install
-            // `--force` reinstalls everything, so an up-to-date tree
-            // must not short-circuit the materialization.
-            && !config.force
-            && let Some(wanted_lockfile) = lockfile
-            && let Some(current) = current_lockfile
-            && wanted_lockfile == current
-            && let Some(modules) = modules_manifest.as_ref()
-            && modules_consistent_with(modules, config, node_linker, included)
-            // A `supportedArchitectures` change alters the skip set
-            // without touching the lockfile or `.modules.yaml`, so the
-            // unchanged-layout premise doesn't hold and the platform
-            // packages must be re-evaluated.
-            && crate::optimistic_repeat_install::recorded_supported_architectures_match(
-                workspace_root,
-                supported_architectures,
-            )
-            // An `allowBuilds` change that now permits a previously-ignored
-            // build must rebuild it, even though the lockfile and layout are
-            // unchanged.
-            && !has_newly_allowed_ignored_builds(modules, config)
-            // The mirror image: an approval the user has since withdrawn
-            // must be re-evaluated, or a strict install would exit 0 on a
-            // package it is no longer allowed to build.
-            && !has_revoked_allowed_builds(modules, config)
-            // A build marker lives in the shared slot, outside every
-            // project-state input checked above. Let materialization inspect
-            // buildable and patched GVS slots instead of declaring the local
-            // tree complete from importer links alone.
-            && !gvs_build_marker_present(wanted_lockfile, config, workspace_root)
-            // An explicit `pacquet rebuild` always re-runs the build phase,
-            // so it never short-circuits here.
-            && rebuild.is_none()
-            && !modules_cache_prune_due
-            && frozen_tree_intact(wanted_lockfile, modules, config, workspace_root, node_linker)
-    {
-        // The full frozen path runs the offline structural
-        // name gate before any materialization; the up-to-date
-        // early return must not skip it (the resolution-verifier
-        // fan-out below is policy-gated and can be empty).
-        pacquet_lockfile_verification::verify_lockfile_dependency_names(wanted_lockfile)
-            .map_err(InstallError::LockfileVerification)?;
-        // Nothing to materialize means no fetch to overlap; verify
-        // eagerly before the up-to-date early return.
-        if let Some(lockfile_verification_override) = lockfile_verification_override {
-            lockfile_verification_override.await.map_err(map_frozen_lockfile_error)?;
-        } else {
-            verify_lockfile_eagerly::<Reporter>(
-                wanted_lockfile,
-                resolution_verifiers,
-                derived_lockfile_path,
-                &config.cache_dir,
-            )
-            .await?;
-        }
-        // Keep `strictDepBuilds` enforced on the up-to-date path: a
-        // rerun after an `ERR_PNPM_IGNORED_BUILDS` failure must not
-        // exit 0 just because the lockfile and layout are unchanged.
-        // Checked after verification (a tampered lockfile fails first)
-        // and before the "up to date" log so the command doesn't
-        // claim success.
-        // `Err` (malformed `allowBuilds`) is unreachable here — the
-        // `has_newly_allowed_ignored_builds` guard above returns `true`
-        // on the same `from_config` error and skips this block — so a
-        // bad policy is surfaced by the full install instead.
-        if config.strict_dep_builds
-            && let Ok(Some(package_names)) = unapproved_recorded_ignored_builds(modules, config)
-        {
-            return Err(InstallError::IgnoredBuilds { package_names });
-        }
-        Reporter::emit(&LogEvent::Pnpm(PnpmLog {
-            level: LogLevel::Info,
-            message: "Lockfile is up to date, resolution step is skipped".to_string(),
-            prefix: prefix.clone(),
-        }));
-        Reporter::emit(&LogEvent::Stage(StageLog {
-            level: LogLevel::Debug,
-            prefix: prefix.clone(),
-            stage: Stage::ImportingDone,
-        }));
-        if (lockfile_synthesized_from_current || lockfile_was_fast_updated) && config.lockfile {
-            wanted_lockfile
-                .save_to_path(&workspace_root.join(Lockfile::FILE_NAME))
-                .map_err(InstallError::SaveWantedLockfile)?;
-        }
-        update_workspace_state(
-            workspace_root,
-            &build_workspace_state(
-                workspace_root,
-                config,
-                node_linker,
-                included,
-                supported_architectures,
-                catalogs,
-                project_manifests,
-                filtered_install,
-            ),
-        )
-        .map_err(InstallError::WriteWorkspaceState)?;
-        Reporter::emit(&LogEvent::Summary(SummaryLog {
-            level: LogLevel::Debug,
-            prefix: prefix.to_owned(),
-        }));
-        return Ok(None);
-    }
-
-    Ok(Some(PreparedModulesState {
-        old_modules,
-        previous_modules_metadata,
-        is_inconsistent,
-        lockfile_verification_override,
-    }))
-}
-
-struct MaterializationInputs<'a, 'install> {
-    tarball_mem_cache: Arc<MemCache>,
-    resolved_packages: &'a ResolvedPackages,
-    http_client: &'a ThrottledClient,
-    http_client_arc: Arc<ThrottledClient>,
-    config: &'static Config,
-    manifest: &'a PackageManifest,
-    lockfile: Option<&'a Lockfile>,
-    take_frozen_path: bool,
-    lockfile_verification_override: Option<super::LockfileVerificationOverride<'install>>,
-    resolution_verifiers: Vec<Arc<dyn ResolutionVerifier>>,
-    derived_lockfile_path: Option<PathBuf>,
-    dependency_groups: Vec<DependencyGroup>,
-    project_manifests: &'a [(PathBuf, &'a PackageManifest)],
-    workspace_projects: Option<&'a [pacquet_workspace::Project]>,
-    requested_importer_ids: Option<&'a HashSet<String>>,
-    real_importer_ids: &'a HashSet<String>,
-    workspace_root: &'a Path,
-    included: IncludedDependencies,
-    node_linker: NodeLinker,
-    rebuild: Option<&'a RebuildOptions>,
-    ignore_manifest_check: bool,
-    mutation: ProjectMutation,
-    current_lockfile: Option<&'a Lockfile>,
-    supported_architectures: Option<&'a pacquet_package_is_installable::SupportedArchitectures>,
-    skip_runtimes: bool,
-    modules_manifest: Option<&'a pacquet_modules_yaml::ModulesLayout>,
-    prior_hoisted_dependencies: Option<&'a HoistedDependencies>,
-    prune_orphans: bool,
-    logged_methods: &'a AtomicU8,
-    update_checksums: bool,
-    meta_cache: Arc<InMemoryPackageMetaCache>,
-    resolve_only: bool,
-    dry_run: bool,
-    can_prompt: bool,
-    update_seed_policy: UpdateSeedPolicy,
-    auth_override: Option<Arc<AuthHeaders>>,
-    resolution_observer: Option<Arc<dyn crate::ResolutionObserver>>,
-    peer_issues_sink: Option<PeerIssuesSink>,
-    deps_requiring_build_sink: Option<DepsRequiringBuildSink>,
-    pnpmfile_hook: Option<Arc<dyn pacquet_hooks::PnpmfileHooks>>,
-    catalogs: &'a Catalogs,
-    prefix: &'a String,
-}
-
-struct MaterializationOutput {
-    ignored_builds: Vec<String>,
-    deferred_builds: Vec<String>,
-    injected_deps: BTreeMap<String, Vec<String>>,
-    hoisted_dependencies: HoistedDependencies,
-    hoisted_locations: BTreeMap<String, Vec<String>>,
-    install_skipped: crate::SkippedSnapshots,
-    fresh_lockfile: Option<Lockfile>,
-}
-
-async fn materialize<Reporter: self::Reporter + 'static>(
-    inputs: MaterializationInputs<'_, '_>,
-) -> Result<MaterializationOutput, InstallError> {
-    let MaterializationInputs {
-        tarball_mem_cache,
-        resolved_packages,
-        http_client,
-        http_client_arc,
-        config,
-        manifest,
-        lockfile,
-        take_frozen_path,
-        lockfile_verification_override,
-        resolution_verifiers,
-        derived_lockfile_path,
-        dependency_groups,
-        project_manifests,
-        workspace_projects,
-        requested_importer_ids,
-        real_importer_ids,
-        workspace_root,
-        included,
-        node_linker,
-        rebuild,
-        ignore_manifest_check,
-        mutation,
-        current_lockfile,
-        supported_architectures,
-        skip_runtimes,
-        modules_manifest,
-        prior_hoisted_dependencies,
-        prune_orphans,
-        logged_methods,
-        update_checksums,
-        meta_cache,
-        resolve_only,
-        dry_run,
-        can_prompt,
-        update_seed_policy,
-        auth_override,
-        resolution_observer,
-        peer_issues_sink,
-        deps_requiring_build_sink,
-        pnpmfile_hook,
-        catalogs,
-        prefix,
-    } = inputs;
-    let ignored_builds: Vec<String>;
-    let deferred_builds: Vec<String>;
-    let injected_deps: BTreeMap<String, Vec<String>>;
-    let effective_node_version =
-        config.node_version.clone().or_else(|| node_version_from_engines_runtime(manifest.value()));
-    let (hoisted_dependencies, hoisted_locations, install_skipped, fresh_lockfile): (
-        HoistedDependencies,
-        BTreeMap<String, Vec<String>>,
-        crate::SkippedSnapshots,
-        Option<Lockfile>,
-    ) = if take_frozen_path {
-        let lockfile = lockfile.expect("dispatch verified lockfile is present");
-        // pnpm's headless installer announces itself whenever it is
-        // entered — also on a cold `node_modules` and on subset
-        // (`--filter`) installs — not only when nothing needs to be
-        // materialized. `pnpm fetch` gets upstream's
-        // ignorePackageManifest wording instead; it is the one
-        // caller combining `ignore_manifest_check` with a non-full
-        // install, and the flag alone can't identify it because
-        // `install --ignore-manifest-check` is a user-facing way to
-        // skip the frozen freshness gate on a full install.
-        // Upstream's headless entry returns before the announcement
-        // for an empty lockfile (`isEmptyLockfile`), and an explicit
-        // `pnpm rebuild` is not an install, so both stay silent.
-        if rebuild.is_none() && !lockfile.is_empty() {
-            let message = if ignore_manifest_check && !mutation.is_full_install() {
-                "Importing packages to virtual store"
-            } else {
-                "Lockfile is up to date, resolution step is skipped"
-            };
-            Reporter::emit(&LogEvent::Pnpm(PnpmLog {
-                level: LogLevel::Info,
-                message: message.to_string(),
-                prefix: prefix.clone(),
-            }));
-        }
-        let initial_materialization_ids = requested_importer_ids.map(|selected| {
-            if matches!(node_linker, NodeLinker::Hoisted) {
-                lockfile.importers.keys().cloned().collect()
-            } else {
-                selected.clone()
-            }
-        });
-        let empty_skipped = crate::SkippedSnapshots::new();
-        let materialization = initial_materialization_ids.as_ref().map(|importer_ids| {
-            crate::materialization_closure(
-                lockfile,
-                workspace_root,
-                importer_ids,
-                included,
-                &empty_skipped,
-            )
-        });
-        let materialization_lockfile =
-            materialization.as_ref().map_or(lockfile, |closure| &closure.lockfile);
-        let project_anchor_ids = match requested_importer_ids {
-            Some(selected) if matches!(node_linker, NodeLinker::Hoisted) => selected.clone(),
-            Some(_) => materialization
-                .as_ref()
-                .expect("selected install has a materialization closure")
-                .importer_ids
-                .clone(),
-            None => real_importer_ids.clone(),
-        };
-        let frozen_project_manifests = project_manifests
-            .iter()
-            .filter(|(project_dir, _)| {
-                let importer_id =
-                    pacquet_workspace::importer_id_from_root_dir(workspace_root, project_dir);
-                project_anchor_ids.contains(&importer_id)
-            })
-            .cloned()
-            .collect::<Vec<_>>();
-        let Lockfile { lockfile_version, importers, packages, snapshots, .. } =
-            materialization_lockfile;
-        let lockfile_major = lockfile_version.major;
-        let supported_lockfile_major = matches!(lockfile_major, 9 | 12);
-        debug_assert!(supported_lockfile_major);
-
-        let mut frozen_verification_override = lockfile_verification_override;
-        if requested_importer_ids.is_some() {
-            if let Some(verification_override) = frozen_verification_override.take() {
-                verification_override.await.map_err(map_frozen_lockfile_error)?;
-            } else {
-                verify_lockfile_eagerly::<Reporter>(
-                    lockfile,
-                    &resolution_verifiers,
-                    derived_lockfile_path.as_deref(),
-                    &config.cache_dir,
-                )
-                .await?;
-            }
-        }
-        let frozen_resolution_verifiers = if requested_importer_ids.is_some() {
-            &[][..]
-        } else {
-            resolution_verifiers.as_slice()
-        };
-
-        let frozen_result = InstallFrozenLockfile {
-            http_client,
-            config,
-            importers,
-            packages: packages.as_ref(),
-            snapshots: snapshots.as_ref(),
-            lockfile: materialization_lockfile,
-            resolution_verifiers: frozen_resolution_verifiers,
-            lockfile_verification_override: frozen_verification_override,
-            lockfile_path: derived_lockfile_path.as_deref(),
-            current_lockfile,
-            // `--force` relinks every package, so the per-snapshot
-            // "unchanged since the previous install" skip must not
-            // see the current lockfile — pnpm's
-            // `lockfileToDepGraph(..., opts.force ? null :
-            // currentLockfile)`. `current_lockfile` itself stays:
-            // pnpm's prune runs on the real current lockfile even
-            // under force.
-            current_snapshots: (!config.force)
-                .then_some(current_lockfile)
-                .flatten()
-                .and_then(|lockfile| lockfile.snapshots.as_ref()),
-            current_packages: (!config.force)
-                .then_some(current_lockfile)
-                .flatten()
-                .and_then(|lockfile| lockfile.packages.as_ref()),
-            dependency_groups,
-            project_manifests: &frozen_project_manifests,
-            package_map_project_manifests: project_manifests,
-            logged_methods,
-            workspace_root,
-            requester: prefix,
-            supported_architectures,
-            skip_runtimes,
-            node_version: effective_node_version.clone(),
-            node_linker,
-            tarball_mem_cache: Some(&tarball_mem_cache),
-            seed_skipped: modules_manifest.map(|manifest| manifest.skipped.clone()),
-            rebuild,
-            prior_hoisted_dependencies,
-            prune_orphans,
-        }
-        .run::<Reporter>()
-        .await
-        // Surface a verification failure as the same top-level
-        // `LockfileVerification` variant the eager paths use, rather
-        // than nesting it under `FrozenLockfile` — the concurrent gate
-        // is the same gate, just run alongside the fetch.
-        .map_err(map_frozen_lockfile_error)?;
-
-        ignored_builds = frozen_result.ignored_builds;
-        deferred_builds = frozen_result.deferred_builds;
-        injected_deps = frozen_result.injected_deps;
-        (
-            frozen_result.hoisted_dependencies,
-            frozen_result.hoisted_locations,
-            frozen_result.skipped,
-            None,
-        )
-    } else {
-        // Re-verify the existing lockfile before the fresh resolve,
-        // matching the pre-resolution gate: a committed lockfile that
-        // bypassed the policy locally is caught here even though the
-        // resolver re-resolves from it. No-op when there's no lockfile
-        // (state 4) or verification is disabled. The fresh path's own
-        // resolution is the slow part, so this stays a blocking gate.
-        if let Some(lockfile_verification_override) = lockfile_verification_override {
-            lockfile_verification_override.await.map_err(map_frozen_lockfile_error)?;
-        } else if let Some(loaded_lockfile) = lockfile {
-            verify_lockfile_eagerly::<Reporter>(
-                loaded_lockfile,
-                &resolution_verifiers,
-                derived_lockfile_path.as_deref(),
-                &config.cache_dir,
-            )
-            .await?;
-        }
-
-        let workspace_packages = build_workspace_packages_map(workspace_projects);
-        // Build the per-importer manifest list. The root importer
-        // (`"."`) always reuses the in-memory `Install.manifest`
-        // — `pacquet add` mutates that value before calling install,
-        // so re-reading from disk would walk the pre-add shape and
-        // miss the freshly-added dep. Sibling importers come from
-        // the `find_workspace_projects` walk, which read them off
-        // disk for `workspace_packages` already.
-        let importer_manifests: BTreeMap<String, &PackageManifest> = project_manifests
-            .iter()
-            .map(|(project_dir, manifest)| {
-                (
-                    pacquet_workspace::importer_id_from_root_dir(workspace_root, project_dir),
-                    *manifest,
-                )
-            })
-            .collect();
-        let fresh_result = InstallWithFreshLockfile {
-            tarball_mem_cache,
-            resolved_packages,
-            http_client,
-            http_client_arc: Arc::clone(&http_client_arc),
-            config,
-            importer_manifests,
-            dependency_groups,
-            logged_methods,
-            requester: prefix,
-            catalogs: catalogs.clone(),
-            lockfile_dir: workspace_root,
-            workspace_packages,
-            update_checksums,
-            meta_cache: Arc::clone(&meta_cache),
-            // States 3 and 4 of the dispatch share this branch.
-            // State 3 (lockfile present but stale or
-            // `preferFrozenLockfile: false`) passes the existing
-            // lockfile so the resolver seeds
-            // `getPreferredVersionsFromLockfileAndManifests` with
-            // already-pinned `(name, version)` pairs — unrelated
-            // entries keep their pins on rewrite (the `update: false`
-            // mode). State 4 (no lockfile) passes `None`.
-            wanted_lockfile: lockfile,
-            node_version: effective_node_version,
-            node_linker,
-            supported_architectures,
-            lockfile_only: resolve_only,
-            skip_runtimes,
-            dry_run,
-            can_prompt,
-            is_full_install: mutation.is_full_install(),
-            update_seed_policy,
-            auth_override,
-            resolution_observer,
-            peer_issues_sink: peer_issues_sink.clone(),
-            deps_requiring_build_sink: deps_requiring_build_sink.as_ref().map(Arc::clone),
-            pnpmfile_hook_override: pnpmfile_hook,
-            real_importer_ids: requested_importer_ids.map(|_| real_importer_ids),
-            selected_importer_ids: requested_importer_ids,
-            current_lockfile,
-            prior_hoisted_dependencies,
-            prune_orphans,
-        }
-        .run::<Reporter>()
-        .await
-        .map_err(InstallError::WithFreshLockfile)?;
-
-        if fresh_result.can_record_lockfile_verification
-            && let Some(lockfile) = fresh_result.wanted_lockfile.as_ref()
-        {
-            // Record under the same path the verification gates key
-            // their cache on, so the next install's stat shortcut hits.
-            let lockfile_path = derived_lockfile_path
-                .clone()
-                .unwrap_or_else(|| workspace_root.join(Lockfile::FILE_NAME));
-            record_lockfile_verified(
-                Some(&config.cache_dir),
-                &lockfile_path,
-                lockfile,
-                &resolution_verifiers,
-            );
-        }
-
-        ignored_builds = fresh_result.ignored_builds;
-        deferred_builds = fresh_result.deferred_builds;
-        injected_deps = fresh_result.injected_deps;
-        (
-            fresh_result.hoisted_dependencies,
-            fresh_result.hoisted_locations,
-            fresh_result.skipped,
-            fresh_result.wanted_lockfile,
-        )
-    };
-
-    Ok(MaterializationOutput {
-        ignored_builds,
-        deferred_builds,
-        injected_deps,
-        hoisted_dependencies,
-        hoisted_locations,
-        install_skipped,
-        fresh_lockfile,
-    })
-}
-
-struct ResolveOnlyCompletionInputs<'a> {
-    resolve_only: bool,
-    dry_run: bool,
-    peer_issues_sink_is_none: bool,
-    existing_wanted_lockfile: Option<&'a Lockfile>,
-    fresh_lockfile: Option<&'a Lockfile>,
-    prefix: &'a String,
-}
-
-fn complete_resolve_only<Reporter: self::Reporter>(
-    inputs: &ResolveOnlyCompletionInputs<'_>,
-) -> Result<bool, InstallError> {
-    if !inputs.resolve_only {
-        return Ok(false);
-    }
-
-    // A sink-driven dry run is a programmatic query, not a CLI preview.
-    if inputs.dry_run && inputs.peer_issues_sink_is_none {
-        use std::io::Write as _;
-        let report =
-            crate::lockfile_diff::render_dry_run_report(&crate::lockfile_diff::diff_lockfiles(
-                inputs.existing_wanted_lockfile,
-                inputs.fresh_lockfile,
-                crate::lockfile_diff::ImporterDiffKey::Specifier,
-            ));
-        let mut stdout = std::io::stdout();
-        let _ = writeln!(stdout, "{report}");
-        let _ = stdout.flush();
-    }
-    Reporter::emit(&LogEvent::Summary(SummaryLog {
-        level: LogLevel::Debug,
-        prefix: inputs.prefix.clone(),
-    }));
-    Ok(true)
-}
-
-struct SelectMaterializedStateInputs<'a> {
-    fresh_lockfile: Option<&'a Lockfile>,
-    loaded_wanted_lockfile: Option<&'a Lockfile>,
-    requested_importer_ids: Option<&'a HashSet<String>>,
-    real_importer_ids: &'a HashSet<String>,
-    workspace_root: &'a Path,
-    included: IncludedDependencies,
-    install_skipped: &'a crate::SkippedSnapshots,
-    node_linker: NodeLinker,
-    current_lockfile: Option<&'a Lockfile>,
-    is_inconsistent: bool,
-    project_manifests: &'a [(PathBuf, &'a PackageManifest)],
-}
-
-struct MaterializedState<'a> {
-    wanted_lockfile: Option<&'a Lockfile>,
-    selected_current_lockfile: Option<Lockfile>,
-    current_lockfile: Option<Lockfile>,
-    project_manifests: Vec<(PathBuf, &'a PackageManifest)>,
-}
-
-fn select_materialized_state<'a>(
-    inputs: &SelectMaterializedStateInputs<'a>,
-) -> MaterializedState<'a> {
-    let wanted_lockfile = inputs.fresh_lockfile.or(inputs.loaded_wanted_lockfile);
-    let selected_current_lockfile = wanted_lockfile.and_then(|wanted| {
-        inputs.requested_importer_ids.map(|requested| {
-            crate::materialization_closure(
-                wanted,
-                inputs.workspace_root,
-                requested,
-                inputs.included,
-                inputs.install_skipped,
-            )
-            .lockfile
-        })
-    });
-    let current_lockfile = wanted_lockfile.map(|wanted| {
-        if inputs.requested_importer_ids.is_some()
-            && matches!(inputs.node_linker, NodeLinker::Hoisted)
-        {
-            crate::filter_lockfile_for_current(wanted, inputs.included, inputs.install_skipped)
-        } else if let Some(requested_importer_ids) = inputs.requested_importer_ids {
-            crate::merge_filtered_current_lockfile(
-                (!inputs.is_inconsistent).then_some(inputs.current_lockfile).flatten(),
-                wanted,
-                requested_importer_ids,
-                inputs.included,
-                inputs.install_skipped,
-                inputs.workspace_root,
-            )
-        } else {
-            crate::filter_lockfile_for_current(wanted, inputs.included, inputs.install_skipped)
-        }
-    });
-    let project_anchor_importer_ids = match inputs.requested_importer_ids {
-        Some(requested) if matches!(inputs.node_linker, NodeLinker::Hoisted) => requested.clone(),
-        Some(requested) => wanted_lockfile.map_or_else(
-            || requested.clone(),
-            |wanted| {
-                crate::materialization_closure(
-                    wanted,
-                    inputs.workspace_root,
-                    requested,
-                    inputs.included,
-                    inputs.install_skipped,
-                )
-                .importer_ids
-            },
-        ),
-        None => inputs.real_importer_ids.clone(),
-    };
-    let project_manifests = inputs
-        .project_manifests
-        .iter()
-        .filter(|(project_dir, _)| {
-            let importer_id =
-                pacquet_workspace::importer_id_from_root_dir(inputs.workspace_root, project_dir);
-            project_anchor_importer_ids.contains(&importer_id)
-        })
-        .cloned()
-        .collect();
-
-    MaterializedState {
-        wanted_lockfile,
-        selected_current_lockfile,
-        current_lockfile,
-        project_manifests,
-    }
-}
-
-struct LinkMaterializedProjectsInputs<'a> {
-    filtered_install: bool,
-    node_linker: NodeLinker,
-    config: &'static Config,
-    current_lockfile: Option<&'a Lockfile>,
-    wanted_lockfile: Option<&'a Lockfile>,
-    workspace_root: &'a Path,
-    project_manifests: &'a [(PathBuf, &'a PackageManifest)],
-    materialized_project_manifests: &'a [(PathBuf, &'a PackageManifest)],
-}
-
-async fn link_materialized_projects<Reporter: self::Reporter + 'static>(
-    inputs: LinkMaterializedProjectsInputs<'_>,
-) -> Result<(), InstallError> {
-    let LinkMaterializedProjectsInputs {
-        filtered_install,
-        node_linker,
-        config,
-        current_lockfile: materialized_current_lockfile,
-        wanted_lockfile,
-        workspace_root,
-        project_manifests,
-        materialized_project_manifests,
-    } = inputs;
-    if filtered_install
-        && !matches!(node_linker, NodeLinker::Hoisted)
-        && crate::should_write_package_map(config, node_linker)
-        && let Some(current) = materialized_current_lockfile.as_ref()
-    {
-        let runtime_major =
-            crate::install_frozen_lockfile::find_runtime_node_major(current.snapshots.as_ref());
-        let configured_major = config
-            .node_version
-            .as_deref()
-            .and_then(crate::install_frozen_lockfile::parse_major_from_version);
-        let engine_name = match runtime_major.or(configured_major) {
-            Some(major) => Some(pacquet_graph_hasher::engine_name(major, None, None)),
-            None if config.enable_global_virtual_store => tokio::task::spawn_blocking(|| {
-                pacquet_graph_hasher::detect_node_major()
-                    .map(|major| pacquet_graph_hasher::engine_name(major, None, None))
-            })
-            .await
-            .ok()
-            .flatten(),
-            None => None,
-        };
-        let allow_build_policy = crate::AllowBuildPolicy::from_config(config)
-            .expect("allow-build policy was validated by the install path");
-        let layout = crate::VirtualStoreLayout::new(
-            config,
-            engine_name.as_deref(),
-            current.snapshots.as_ref(),
-            current.packages.as_ref(),
-            Some(&allow_build_policy),
-            Some(workspace_root),
-        );
-        crate::package_map::write_package_map(
-            current,
-            &crate::package_map::PackageMapOptions {
-                lockfile_dir: workspace_root,
-                modules_dir: &config.modules_dir,
-                package_map_type: config.node_package_map_type,
-                layout: &layout,
-                project_manifests,
-            },
-        )
-        .map_err(InstallError::WritePackageMap)?;
-    }
-
-    // Materialize `link:` direct deps straight from the in-memory
-    // project manifests. `excludeLinksFromLockfile` keeps them out
-    // of the lockfile importers, so the lockfile-driven symlink
-    // passes inside the frozen/fresh paths never see them; pnpm
-    // v11's `linkDirectDeps` linked them from the projects
-    // regardless. Aliases the wanted lockfile *does* track are
-    // skipped — those belong to the lockfile passes (and their
-    // dedupe decisions). See [`crate::link_manifest_link_deps`].
-    // These are importer symlinks like any other, so
-    // `virtualStoreOnly` skips them too.
-    if !config.virtual_store_only {
-        crate::link_manifest_link_deps::<Reporter>(
-            workspace_root,
-            materialized_project_manifests,
-            wanted_lockfile.and_then(|lockfile| {
-                (!lockfile.importers.is_empty()).then_some(&lockfile.importers)
-            }),
-            // Honor a `modulesDir` override the same way the
-            // lockfile-driven symlink pass does.
-            config.modules_dir.file_name().unwrap_or_else(|| std::ffi::OsStr::new("node_modules")),
-            &crate::shim_extra_node_paths(config, node_linker),
-        )
-        .map_err(InstallError::LinkManifestLinkDeps)?;
-    }
-
-    Ok(())
-}
-
-struct CommitModulesStateInputs<'a> {
-    prior_modules: Option<&'a pacquet_modules_yaml::ModulesLayout>,
-    config: &'static Config,
-    workspace_root: &'a Path,
-    materialized_current_lockfile: Option<&'a Lockfile>,
-    selected_current_lockfile: Option<&'a Lockfile>,
-    materialized_project_manifests: &'a [(PathBuf, &'a PackageManifest)],
-    included: IncludedDependencies,
-    install_skipped: &'a crate::SkippedSnapshots,
-    node_linker: NodeLinker,
-    filtered_install: bool,
-    is_inconsistent: bool,
-    previous_modules_metadata: Option<&'a Modules>,
-    hoisted_dependencies: HoistedDependencies,
-    hoisted_locations: BTreeMap<String, Vec<String>>,
-    injected_deps: BTreeMap<String, Vec<String>>,
-    ignored_builds: &'a [String],
-    deferred_builds: Vec<String>,
-    rebuild: Option<&'a RebuildOptions>,
-    take_frozen_path: bool,
-    lockfile_synthesized_from_current: bool,
-    lockfile_was_fast_updated: bool,
-    loaded_wanted_lockfile: Option<&'a Lockfile>,
-}
-
-fn commit_modules_state(inputs: CommitModulesStateInputs<'_>) -> Result<(), InstallError> {
-    let CommitModulesStateInputs {
-        prior_modules,
-        config,
-        workspace_root,
-        materialized_current_lockfile,
-        selected_current_lockfile,
-        materialized_project_manifests,
-        included,
-        install_skipped,
-        node_linker,
-        filtered_install,
-        is_inconsistent,
-        previous_modules_metadata,
-        hoisted_dependencies,
-        hoisted_locations,
-        injected_deps,
-        ignored_builds,
-        deferred_builds,
-        rebuild,
-        take_frozen_path,
-        lockfile_synthesized_from_current,
-        lockfile_was_fast_updated,
-        loaded_wanted_lockfile,
-    } = inputs;
-    let now = SystemTime::now();
-    let effective_virtual_store_dir = config.effective_virtual_store_dir();
-    // Decide "this is the global store" from the resolved paths, not
-    // the `enableGlobalVirtualStore` flag alone: the global store is
-    // shared across projects, so a config that points `virtualStoreDir`
-    // at it must not be pruned even when the flag is off.
-    let is_global_virtual_store = crate::prune_virtual_store::same_dir(
-        effective_virtual_store_dir,
-        &config.global_virtual_store_dir,
-    );
-    // `did_prune` tracks whether the sweep actually ran (enumerated the
-    // store), not just whether the throttle allowed it. It stays false
-    // when there is no wanted lockfile to derive the needed set from
-    // (e.g. `config.lockfile == false` leaves both `fresh_lockfile` and
-    // a loaded `lockfile` absent), when the target is refused as unsafe,
-    // or when enumeration failed. `prunedAt` must not advance on a run
-    // where nothing was swept, or the next real sweep is throttled off
-    // for `modulesCacheMaxAge`.
-    let did_prune = if crate::prune_virtual_store::should_prune_virtual_store(
-        is_global_virtual_store,
-        prior_modules.as_ref().map(|modules| modules.pruned_at.as_str()),
-        config.modules_cache_max_age,
-        now,
-    ) {
-        match materialized_current_lockfile.as_ref() {
-            // Sweep the canonicalized prune target returned by the
-            // containment check, never the raw configured path: deleting
-            // from the validated path closes the time-of-check/time-of-use
-            // gap a symlink swap would otherwise open.
-            Some(wanted) => {
-                if let Some(prune_dir) = crate::prune_virtual_store::prune_target_within_modules(
-                    effective_virtual_store_dir,
-                    &config.modules_dir,
-                ) {
-                    crate::prune_virtual_store::prune_virtual_store(
-                        &prune_dir,
-                        wanted.snapshots.iter().flat_map(|snapshots| snapshots.keys()),
-                        install_skipped,
-                        config.virtual_store_dir_max_length as usize,
-                    )
-                    .is_some()
-                } else {
-                    // A wanted lockfile exists but the store path is unsafe
-                    // (escapes node_modules); refuse the destructive sweep.
-                    tracing::warn!(
-                        virtual_store_dir = %effective_virtual_store_dir.display(),
-                        modules_dir = %config.modules_dir.display(),
-                        "skipping virtual-store prune: the virtual store is not inside node_modules",
-                    );
-                    false
-                }
-            }
-            None => false,
-        }
-    } else {
-        false
-    };
-
-    // Stamp `prunedAt` only when the sweep ran (or there was no prior
-    // `.modules.yaml`); otherwise preserve the recorded timestamp so
-    // the throttle keeps counting from the last real prune.
-    let pruned_at = match (&prior_modules, did_prune) {
-        (Some(prior), false) => prior.pruned_at.clone(),
-        _ => httpdate::fmt_http_date(now),
-    };
-
-    // The projects whose own install scripts `--ignore-scripts`
-    // skipped are owed a build just like the dependencies the build
-    // phase deferred, and are recorded by importer id.
-    let deferred_projects = config.ignore_scripts.then(|| {
-        materialized_project_manifests
-            .iter()
-            .filter(|(project_dir, manifest)| {
-                project_requires_lifecycle_scripts(project_dir, manifest)
-            })
-            .map(|(project_dir, _)| {
-                pacquet_workspace::importer_id_from_root_dir(workspace_root, project_dir)
-            })
-            .collect::<Vec<_>>()
-    });
-    let previous_pending_builds =
-        prior_modules.map_or(&[][..], |modules| modules.pending_builds.as_slice());
-    // The build phase settles a dependency only when it actually
-    // rebuilt it, so a `pnpm rebuild --pending` that the policy still
-    // blocks (`allowBuilds: None`/`false`) leaves the debt in place.
-    // Reuse the same policy `BuildModules` ran under; on a rebuild a
-    // selected, approved dependency always runs (force-rebuild
-    // bypasses the side-effects cache gate), so policy approval is a
-    // faithful stand-in for "was rebuilt".
-    let rebuild_build_policy =
-        rebuild.as_ref().and_then(|_| crate::AllowBuildPolicy::from_config(config).ok());
-    let pending_builds = merge_pending_builds(
-        previous_pending_builds,
-        deferred_projects.into_iter().flatten().chain(deferred_builds),
-        materialized_current_lockfile,
-        rebuild,
-        rebuild_build_policy.as_ref(),
-    );
-
-    let mut next_modules = build_modules_manifest(
-        config,
-        node_linker,
-        included,
-        hoisted_dependencies,
-        hoisted_locations,
-        injected_deps,
-        install_skipped,
-        ignored_builds,
-        pending_builds,
-        pruned_at,
-    );
-    if filtered_install
-        && !matches!(node_linker, NodeLinker::Hoisted)
-        && !is_inconsistent
-        && let (Some(previous), Some(current), Some(selected)) = (
-            previous_modules_metadata.as_ref(),
-            materialized_current_lockfile.as_ref(),
-            selected_current_lockfile.as_ref(),
-        )
-    {
-        merge_filtered_modules_metadata(&mut next_modules, previous, current, selected);
-    }
-    write_modules_manifest::<Host>(&config.modules_dir, next_modules)
-        .map_err(InstallError::WriteModules)?;
-
-    // Write `<virtual_store_dir>/lock.yaml`. Captures what was
-    // actually materialized so the next install can diff each
-    // snapshot against it and skip the unchanged
-    // slots. Persist *after* `write_modules_manifest` succeeds so
-    // a manifest failure can't leave a fresh current-lockfile
-    // pointing at incomplete install state — the next frozen
-    // reinstall would otherwise diff against a graph that never
-    // finished committing (review on <https://github.com/pnpm/pacquet/pull/442>).
-    //
-    // A filtered isolated/PnP install merges its newly materialized
-    // closure into compatible prior current state, while a hoisted
-    // install records the full shared graph it materialized. This
-    // keeps the file aligned with physical state without discarding
-    // unselected slots that remain on disk.
-    if let Some(lockfile) = materialized_current_lockfile.as_ref() {
-        // Filter the wanted lockfile down to the snapshots that
-        // were actually materialized: dep maps the user excluded
-        // (`--no-optional`, `--no-dev`) plus snapshots the
-        // install-time skip set transiently dropped (a fetch
-        // failure, `--no-optional`-only entries). The next install
-        // diffs against this filtered shape so dropped snapshots
-        // aren't mistaken for already-done work.
-        lockfile
-            .save_current_to_virtual_store_dir(&config.virtual_store_dir)
-            .map_err(InstallError::SaveCurrentLockfile)?;
-    }
-
-    // Regenerate `pnpm-lock.yaml` from the synthesized snapshot when
-    // the wanted lockfile was reconstructed from
-    // `<virtual_store_dir>/lock.yaml`. The no-op short-circuit above
-    // handles the common case; this branch covers the rare path where
-    // `.modules.yaml` was wiped or inconsistent and the frozen install
-    // had to relink.
-    if take_frozen_path
-        && (lockfile_synthesized_from_current || lockfile_was_fast_updated)
-        && config.lockfile
-        && let Some(updated) = loaded_wanted_lockfile
-    {
-        updated
-            .save_to_path(&workspace_root.join(Lockfile::FILE_NAME))
-            .map_err(InstallError::SaveWantedLockfile)?;
-    }
-
-    Ok(())
-}
-
-#[derive(Clone, Copy)]
-struct MaterializedProjectScriptsInputs<'a, 'selection> {
-    config: &'static Config,
-    node_linker: NodeLinker,
-    workspace_root: &'a Path,
-    rebuild: Option<&'a RebuildOptions>,
-    mutation: ProjectMutation,
-    manifest_dir: &'a Path,
-    selection: Option<&'a WorkspaceInstallSelection<'selection>>,
-    project_manifests: &'a [(PathBuf, &'a PackageManifest)],
-    materialized_project_manifests: &'a [(PathBuf, &'a PackageManifest)],
-    materialized_current_lockfile: Option<&'a Lockfile>,
-}
-
-fn run_materialized_project_scripts<Reporter: self::Reporter>(
-    inputs: MaterializedProjectScriptsInputs<'_, '_>,
-) -> Result<(), InstallError> {
-    let MaterializedProjectScriptsInputs {
-        config,
-        node_linker,
-        workspace_root,
-        rebuild,
-        mutation,
-        manifest_dir,
-        selection,
-        project_manifests,
-        materialized_project_manifests,
-        materialized_current_lockfile,
-    } = inputs;
-    // A `pnpm rebuild --pending` is the exception to the mutation
-    // gate: it installs no manifest, but the projects it names are
-    // exactly the ones whose scripts an earlier `--ignore-scripts`
-    // install deferred, and running them is what lets the install
-    // drop those entries from `pendingBuilds` (see
-    // [`merge_pending_builds`]).
-    let projects_to_run: Vec<(std::path::PathBuf, &PackageManifest)> =
-        if config.ignore_scripts || config.virtual_store_only {
-            Vec::new()
-        } else if let Some(rebuild) = rebuild {
-            materialized_project_manifests
-                .iter()
-                .filter(|(project_dir, _)| {
-                    let importer_id =
-                        pacquet_workspace::importer_id_from_root_dir(workspace_root, project_dir);
-                    rebuild.pending_projects.contains(&importer_id)
-                })
-                .cloned()
-                .collect()
-        } else {
-            projects_running_own_scripts(&ProjectScriptsInputs {
-                mutation,
-                workspace_root,
-                active_project_dir: manifest_dir,
-                selected_dirs: selection.map(|selection| selection.selected_dirs),
-                project_manifests,
-                materialized_project_manifests,
-            })
-        };
-    if !projects_to_run.is_empty() {
-        let project_groups = order_project_lifecycle_groups(
-            &projects_to_run,
-            selection.map(|selection| selection.ordered_groups),
-            workspace_root,
-            materialized_current_lockfile,
-        )?;
-        if !project_groups.is_empty() {
-            run_projects_lifecycle_scripts::<Reporter>(
-                &project_groups,
-                config,
-                node_linker,
-                workspace_root,
-            )?;
-        }
-        if let Some(rebuild) = rebuild {
-            drain_settled_projects::<Host>(&config.modules_dir, &rebuild.pending_projects)?;
-        }
-    }
-
-    Ok(())
-}
-
-struct ReportInstallCompletionInputs<'a> {
-    config: &'static Config,
-    workspace_root: &'a Path,
-    prefix: String,
-    ignored_builds: Vec<String>,
-}
-
-fn report_install_completion<Reporter: self::Reporter>(
-    inputs: ReportInstallCompletionInputs<'_>,
-) -> Result<(), InstallError> {
-    let ReportInstallCompletionInputs { config, workspace_root, prefix, ignored_builds } = inputs;
-    // `pnpm:summary` closes the install and lets the reporter render
-    // the accumulated `pnpm:root` events as a "+N -M" block. Must
-    // come after `importing_done`.
-    Reporter::emit(&LogEvent::Summary(SummaryLog { level: LogLevel::Debug, prefix }));
-
-    // A global install is exempt from the scaffold below: its root is a
-    // throwaway per-group directory, and the approval prompt that
-    // follows it records the ignored builds against the stable global
-    // packages dir instead.
-    let is_global_install = config
-        .global_pkg_dir
-        .as_deref()
-        .is_some_and(|global_pkg_dir| workspace_root.starts_with(global_pkg_dir));
-    // Leave the user a line to edit in `pnpm-workspace.yaml` for every
-    // build this install blocked, so approving one is an edit rather
-    // than recalling the `allowBuilds` shape. Written before the strict
-    // failure below, which is the very run whose message it answers.
-    if !ignored_builds.is_empty() && !is_global_install {
-        let allow_build_keys: BTreeSet<String> = ignored_builds
-            .iter()
-            .map(|dep_path| crate::allow_build_key_from_ignored_build(dep_path))
-            .collect();
-        pacquet_workspace_manifest_writer::scaffold_allow_builds(
-            config.workspace_dir.as_deref().unwrap_or(workspace_root),
-            allow_build_keys.iter().map(String::as_str),
-        )
-        .map_err(InstallError::ScaffoldAllowBuilds)?;
-    }
-
-    // When `strictDepBuilds` is on (the default), an install that
-    // blocked any dependency build script fails with
-    // `ERR_PNPM_IGNORED_BUILDS` *after* the artifacts are written, so
-    // the package is still added/installed and the user approves the
-    // builds and reinstalls.
-    if config.strict_dep_builds && !ignored_builds.is_empty() {
-        return Err(InstallError::IgnoredBuilds { package_names: ignored_builds });
-    }
-
-    Ok(())
-}
-
-struct ApplyMaterializationInputs<'a, 'selection> {
-    resolve_only: bool,
-    dry_run: bool,
-    peer_issues_sink_is_none: bool,
-    existing_wanted_lockfile: Option<&'a Lockfile>,
-    fresh_lockfile: Option<Lockfile>,
-    prefix: String,
-    lockfile: Option<&'a Lockfile>,
-    requested_importer_ids: Option<HashSet<String>>,
-    workspace_root: PathBuf,
-    included: IncludedDependencies,
-    install_skipped: crate::SkippedSnapshots,
-    node_linker: NodeLinker,
-    current_lockfile: Option<Lockfile>,
-    real_importer_ids: HashSet<String>,
-    project_manifests: Vec<(PathBuf, &'a PackageManifest)>,
-    filtered_install: bool,
-    is_inconsistent: bool,
-    previous_modules_metadata: Option<Modules>,
-    config: &'static Config,
-    hoisted_dependencies: HoistedDependencies,
-    hoisted_locations: BTreeMap<String, Vec<String>>,
-    injected_deps: BTreeMap<String, Vec<String>>,
-    ignored_builds: Vec<String>,
-    deferred_builds: Vec<String>,
-    modules_manifest: Option<pacquet_modules_yaml::ModulesLayout>,
-    rebuild: Option<RebuildOptions>,
-    take_frozen_path: bool,
-    lockfile_synthesized_from_current: bool,
-    lockfile_was_fast_updated: bool,
-    mutation: ProjectMutation,
-    manifest_dir: &'a Path,
-    selection: Option<WorkspaceInstallSelection<'selection>>,
-    supported_architectures: Option<pacquet_package_is_installable::SupportedArchitectures>,
-    catalogs: Catalogs,
-}
-
-async fn apply_materialization_result<Reporter: self::Reporter + 'static>(
-    inputs: ApplyMaterializationInputs<'_, '_>,
-) -> Result<(), InstallError> {
-    let ApplyMaterializationInputs {
-        resolve_only,
-        dry_run,
-        peer_issues_sink_is_none,
-        existing_wanted_lockfile,
-        fresh_lockfile,
-        prefix,
-        lockfile,
-        requested_importer_ids,
-        workspace_root,
-        included,
-        install_skipped,
-        node_linker,
-        current_lockfile,
-        real_importer_ids,
-        project_manifests,
-        filtered_install,
-        is_inconsistent,
-        previous_modules_metadata,
-        config,
-        hoisted_dependencies,
-        hoisted_locations,
-        injected_deps,
-        ignored_builds,
-        deferred_builds,
-        modules_manifest,
-        rebuild,
-        take_frozen_path,
-        lockfile_synthesized_from_current,
-        lockfile_was_fast_updated,
-        mutation,
-        manifest_dir,
-        selection,
-        supported_architectures,
-        catalogs,
-    } = inputs;
-    let modules_manifest = modules_manifest.as_ref();
-    tracing::info!(target: "pacquet::install", "Complete all");
-
-    if complete_resolve_only::<Reporter>(&ResolveOnlyCompletionInputs {
-        resolve_only,
-        dry_run,
-        peer_issues_sink_is_none,
-        existing_wanted_lockfile,
-        fresh_lockfile: fresh_lockfile.as_ref(),
-        prefix: &prefix,
-    })? {
-        return Ok(());
-    }
-
-    let MaterializedState {
-        wanted_lockfile: materialized_wanted_lockfile,
-        selected_current_lockfile,
-        current_lockfile: materialized_current_lockfile,
-        project_manifests: materialized_project_manifests,
-    } = select_materialized_state(&SelectMaterializedStateInputs {
-        fresh_lockfile: fresh_lockfile.as_ref(),
-        loaded_wanted_lockfile: lockfile,
-        requested_importer_ids: requested_importer_ids.as_ref(),
-        real_importer_ids: &real_importer_ids,
-        workspace_root: &workspace_root,
-        included,
-        install_skipped: &install_skipped,
-        node_linker,
-        current_lockfile: current_lockfile.as_ref(),
-        is_inconsistent,
-        project_manifests: &project_manifests,
-    });
-
-    link_materialized_projects::<Reporter>(LinkMaterializedProjectsInputs {
-        filtered_install,
-        node_linker,
-        config,
-        current_lockfile: materialized_current_lockfile.as_ref(),
-        wanted_lockfile: materialized_wanted_lockfile,
-        workspace_root: &workspace_root,
-        project_manifests: &project_manifests,
-        materialized_project_manifests: &materialized_project_manifests,
-    })
-    .await?;
-
-    commit_modules_state(CommitModulesStateInputs {
-        prior_modules: modules_manifest,
-        config,
-        workspace_root: &workspace_root,
-        materialized_current_lockfile: materialized_current_lockfile.as_ref(),
-        selected_current_lockfile: selected_current_lockfile.as_ref(),
-        materialized_project_manifests: &materialized_project_manifests,
-        included,
-        install_skipped: &install_skipped,
-        node_linker,
-        filtered_install,
-        is_inconsistent,
-        previous_modules_metadata: previous_modules_metadata.as_ref(),
-        hoisted_dependencies,
-        hoisted_locations,
-        injected_deps,
-        ignored_builds: &ignored_builds,
-        deferred_builds,
-        rebuild: rebuild.as_ref(),
-        take_frozen_path,
-        lockfile_synthesized_from_current,
-        lockfile_was_fast_updated,
-        loaded_wanted_lockfile: lockfile,
-    })?;
-
-    run_materialized_project_scripts::<Reporter>(MaterializedProjectScriptsInputs {
-        config,
-        node_linker,
-        workspace_root: &workspace_root,
-        rebuild: rebuild.as_ref(),
-        mutation,
-        manifest_dir,
-        selection: selection.as_ref(),
-        project_manifests: &project_manifests,
-        materialized_project_manifests: &materialized_project_manifests,
-        materialized_current_lockfile: materialized_current_lockfile.as_ref(),
-    })?;
-
-    // Write `node_modules/.pnpm-workspace-state-v1.json`.
-    // pnpm's `verifyDepsBeforeRun` gate bails to "outdated" the
-    // moment this file is missing, forcing `pnpm install` to rerun.
-    // Writing it after both the `.modules.yaml` and the current
-    // lockfile succeed keeps the file pointing at a fully committed
-    // install.
-    update_workspace_state(
-        &workspace_root,
-        &build_workspace_state(
-            &workspace_root,
-            config,
-            node_linker,
-            included,
-            supported_architectures.as_ref(),
-            &catalogs,
-            &project_manifests,
-            filtered_install,
-        ),
-    )
-    .map_err(InstallError::WriteWorkspaceState)?;
-
-    report_install_completion::<Reporter>(ReportInstallCompletionInputs {
-        config,
-        workspace_root: &workspace_root,
-        prefix,
-        ignored_builds,
-    })
 }
